Set up Publishing to npm
Adds a npm run build task which compiles CSS + JS to the `dist` folder.
The CSS generates minified and unminified versions. The JS generates UMD
bundles (can be used in browser via AMD or directly using globals, and
can be used in node via commonjs), both minified and unminified, and it
generates an unminified ESM bundle.