Add Changelog (#1329)
Why didn't I add a changelog back in 2020 when I set this up? Good
question. I'm as annoyed at past Scott as you are, since I had to find
an app to generate a changelog file from our git history, and do a lot
of manual cleanup on it. Either way, here it is now, for future use.

Fixes #273