Replace Storybook with Eleventy (#1337)
This PR basically rebuilds the pattern library site (not the pattern
library CSS). Previously, we were using Storybook, which is a powerful,
full-featured, React-based tool that I've used professionally with some
success. What I did not count on was that Storybook would be releasing
major revisions on a regular cadence, and that upgrading would prove to
be a nightmare. Over time, doing work on MLTSHP has become something I
dread, rather than something I look forward to, which is a real problem,
since I'm the primary maintainer of the pattern library.

So, I decided to throw out Storybook entirely and replace it with a
simpler static site generated with Eleventy, a product that is joyfully
easy to build, with an emphasis on speed and a low number of
dependencies. I use Eleventy for dozens of sites both professionally
and personally, and maintenance and major version upgrades are easy
to handle.

When I mentioned this upgrade to @bradchoate he expressed concern at
losing some of the genuinely nice features of Storybook, including the
ability to toggle between light and dark mode, and to resize the preview
windows to see what it looks like on small screens. (Also, a11y tests,
but that'll be covered in a followup PR.)

So I've made sure to retain those features, with simpler vanilla
JS-based functionality. There's a theme toggle button in the primary
navigation, and any patterns that depend on viewport size (typically
full-page previews) are displayed in a resizable iframe, and have a set
of buttons to easily resize to our responsive breakpoints.

All in all, I'm extremely pleased with this change, and while it's a
beast of a PR to review, I think it sets the pattern library up to be
dramatically easier to maintain moving forward, and hopefully make it
easier for others to contribute.